refactor(paths): fix paths and webpack config
