+++ package.json
@@ -10,7 +10,7 @@
     "lint-es6": "eslint --ext .js --ext .jsx ./app ./targets ./test --ignore-path targets/xod-espruino/runtime.js || true",
     "lint-espruino": "eslint --config .eslintrc.espruino.js targets/xod-espruino/runtime.js nodes/js nodes/espruino || true",
     "lint": "npm-run-all lint-es6 lint-espruino",
-    "test": "mocha-webpack \"test/**/*Spec.js\"",
+    "test": "mocha-webpack \"test/**/*.spec.js\"",
     "travis": "npm-run-all --continue-on-error lint test"
   },
   "repository": {},
