+++ src/utils/espruino.js
@@ -1,7 +1,7 @@
 
-import { upload } from 'xod-espruino/upload';
-import transpile from 'xod-espruino/transpiler';
-import runtime from '!raw!xod-espruino/runtime';
+import { upload } from '../runtime/xod-espruino/upload';
+import transpile from '../runtime/xod-espruino/transpiler';
+import runtime from '!raw!../runtime/xod-espruino/runtime';
 
 export function uploadToEspruino(project, progress) {
   const code = transpile({ project, runtime });
