+++ test/selector.spec.js
@@ -1,10 +1,10 @@
 import R from 'ramda';
 import chai from 'chai';
-import initialState from '../app/state';
-import Selectors from '../app/selectors';
-import * as PIN_TYPE from '../app/constants/pinType';
-import * as PIN_DIRECTION from '../app/constants/pinDirection';
-import { LINK_ERRORS } from '../app/constants/errorMessages';
+import initialState from '../src/client/state';
+import Selectors from '../src/client/selectors';
+import * as PIN_TYPE from '../src/client/constants/pinType';
+import * as PIN_DIRECTION from '../src/client/constants/pinDirection';
+import { LINK_ERRORS } from '../src/client/constants/errorMessages';
 
 describe('Link selector', () => {
   describe('while validating link creating', () => {
